+++ nbproject/project.properties
@@ -164,7 +164,7 @@ main.class=Interfaces.Cart
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
-platform.active=JDK_13
+platform.active=default_platform
 run.classpath=\
     ${javac.classpath}:\
     ${build.classes.dir}
