+++ src/Project/BarBilling.java
@@ -1,0 +1,722 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package Project;
+
+import Table.TableModel.LiquorInventory;
+import java.sql.ResultSet;
+import java.text.NumberFormat;
+import javax.swing.JOptionPane;
+import javax.swing.table.DefaultTableModel;
+import net.proteanit.sql.DbUtils;
+import SQL.DBHelper;
+
+/**
+ *
+ * @author Menaka
+ */
+public class BarBilling extends javax.swing.JPanel {
+    
+
+    /**
+     * Creates new form BarBilling
+     */
+    public BarBilling() {
+        initComponents();
+        ShowInventory();
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        Left = new javax.swing.JPanel();
+        jButton1 = new javax.swing.JButton();
+        jButton2 = new javax.swing.JButton();
+        jButton3 = new javax.swing.JButton();
+        jButton4 = new javax.swing.JButton();
+        jPanel2 = new javax.swing.JPanel();
+        jLabel3 = new javax.swing.JLabel();
+        jTextField1 = new javax.swing.JTextField();
+        chckbtn = new javax.swing.JButton();
+        jScrollPane1 = new javax.swing.JScrollPane();
+        liquortable = new javax.swing.JTable();
+        jLabel10 = new javax.swing.JLabel();
+        jSeparator1 = new javax.swing.JSeparator();
+        jLabel5 = new javax.swing.JLabel();
+        jLabel6 = new javax.swing.JLabel();
+        iid = new javax.swing.JTextField();
+        quanty = new javax.swing.JTextField();
+        addToTable = new javax.swing.JButton();
+        jButton5 = new javax.swing.JButton();
+        jScrollPane2 = new javax.swing.JScrollPane();
+        customerbill = new javax.swing.JTable();
+        jLabel4 = new javax.swing.JLabel();
+        jLabel15 = new javax.swing.JLabel();
+        jLabel7 = new javax.swing.JLabel();
+        jLabel8 = new javax.swing.JLabel();
+        txtbillid = new javax.swing.JLabel();
+        txttot = new javax.swing.JLabel();
+        disfield = new javax.swing.JTextField();
+        jSeparator4 = new javax.swing.JSeparator();
+        txtnet = new javax.swing.JLabel();
+        jSeparator6 = new javax.swing.JSeparator();
+        jLabel9 = new javax.swing.JLabel();
+        calcbill = new javax.swing.JButton();
+        jButton6 = new javax.swing.JButton();
+        jSeparator7 = new javax.swing.JSeparator();
+        btnUpdate = new javax.swing.JButton();
+        btnDelete = new javax.swing.JButton();
+        AddToBill = new javax.swing.JButton();
+        temp_bar_holder = new javax.swing.JLabel();
+        jLabel11 = new javax.swing.JLabel();
+
+        setMaximumSize(new java.awt.Dimension(1360, 590));
+        setMinimumSize(new java.awt.Dimension(1360, 590));
+        setPreferredSize(new java.awt.Dimension(1360, 590));
+
+        Left.setBackground(new java.awt.Color(51, 51, 51));
+        Left.setMaximumSize(new java.awt.Dimension(235, 590));
+        Left.setMinimumSize(new java.awt.Dimension(235, 590));
+        Left.setPreferredSize(new java.awt.Dimension(235, 590));
+
+        jButton1.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/Group 33 Black.png"))); // NOI18N
+        jButton1.setMaximumSize(new java.awt.Dimension(180, 47));
+        jButton1.setMinimumSize(new java.awt.Dimension(180, 47));
+        jButton1.setPreferredSize(new java.awt.Dimension(180, 47));
+        jButton1.setRolloverIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/Group 73 Orange.png"))); // NOI18N
+        jButton1.setRolloverSelectedIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/Group 73 Orange.png"))); // NOI18N
+        jButton1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton1ActionPerformed(evt);
+            }
+        });
+
+        jButton2.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/Group 34 Black.png"))); // NOI18N
+        jButton2.setMaximumSize(new java.awt.Dimension(180, 47));
+        jButton2.setMinimumSize(new java.awt.Dimension(180, 47));
+        jButton2.setPreferredSize(new java.awt.Dimension(180, 47));
+        jButton2.setRolloverIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/Group 74 Orange.png"))); // NOI18N
+
+        jButton3.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/Group 35 Black.png"))); // NOI18N
+        jButton3.setMaximumSize(new java.awt.Dimension(180, 47));
+        jButton3.setMinimumSize(new java.awt.Dimension(180, 47));
+        jButton3.setPreferredSize(new java.awt.Dimension(180, 47));
+        jButton3.setRolloverIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/Group 75 Orange.png"))); // NOI18N
+        jButton3.setRolloverSelectedIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/Group 75 Orange.png"))); // NOI18N
+
+        jButton4.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/Homebtn.png"))); // NOI18N
+        jButton4.setMaximumSize(new java.awt.Dimension(178, 45));
+        jButton4.setMinimumSize(new java.awt.Dimension(178, 45));
+        jButton4.setPreferredSize(new java.awt.Dimension(178, 45));
+
+        javax.swing.GroupLayout LeftLayout = new javax.swing.GroupLayout(Left);
+        Left.setLayout(LeftLayout);
+        LeftLayout.setHorizontalGroup(
+            LeftLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(LeftLayout.createSequentialGroup()
+                .addGap(19, 19, 19)
+                .addGroup(LeftLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                    .addGroup(LeftLayout.createSequentialGroup()
+                        .addGap(2, 2, 2)
+                        .addComponent(jButton4, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                    .addGroup(LeftLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(jButton3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(jButton2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(jButton1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                .addGap(36, 36, 36))
+        );
+        LeftLayout.setVerticalGroup(
+            LeftLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(LeftLayout.createSequentialGroup()
+                .addGap(130, 130, 130)
+                .addComponent(jButton1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(30, 30, 30)
+                .addComponent(jButton2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(32, 32, 32)
+                .addComponent(jButton3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 173, Short.MAX_VALUE)
+                .addComponent(jButton4, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(39, 39, 39))
+        );
+
+        jPanel2.setMaximumSize(new java.awt.Dimension(1125, 590));
+        jPanel2.setMinimumSize(new java.awt.Dimension(1125, 590));
+        jPanel2.setPreferredSize(new java.awt.Dimension(1125, 590));
+
+        jLabel3.setFont(new java.awt.Font("Tahoma", 0, 18)); // NOI18N
+        jLabel3.setText("Booking ID");
+
+        jTextField1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jTextField1ActionPerformed(evt);
+            }
+        });
+
+        chckbtn.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        chckbtn.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/chklite.png"))); // NOI18N
+        chckbtn.setMaximumSize(new java.awt.Dimension(180, 47));
+        chckbtn.setMinimumSize(new java.awt.Dimension(180, 47));
+        chckbtn.setPreferredSize(new java.awt.Dimension(180, 47));
+        chckbtn.setRolloverIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/chkdark.png"))); // NOI18N
+        chckbtn.setRolloverSelectedIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/chkdark.png"))); // NOI18N
+        chckbtn.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                chckbtnActionPerformed(evt);
+            }
+        });
+
+        liquortable.setModel(new javax.swing.table.DefaultTableModel(
+            new Object [][] {
+
+            },
+            new String [] {
+                "Item ID", "Name", "Size", "Brand", "No of Stocks", "Unit Price"
+            }
+        ) {
+            Class[] types = new Class [] {
+                java.lang.Object.class, java.lang.Object.class, java.lang.Object.class, java.lang.Object.class, java.lang.Object.class, java.lang.Float.class
+            };
+
+            public Class getColumnClass(int columnIndex) {
+                return types [columnIndex];
+            }
+        });
+        jScrollPane1.setViewportView(liquortable);
+
+        jLabel10.setFont(new java.awt.Font("Tahoma", 1, 20)); // NOI18N
+        jLabel10.setText("Customer Check");
+
+        jLabel5.setFont(new java.awt.Font("Tahoma", 0, 18)); // NOI18N
+        jLabel5.setText("Item ID");
+
+        jLabel6.setFont(new java.awt.Font("Tahoma", 0, 18)); // NOI18N
+        jLabel6.setText("Quantity");
+
+        iid.setToolTipText("");
+        iid.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                iidActionPerformed(evt);
+            }
+        });
+
+        quanty.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                quantyActionPerformed(evt);
+            }
+        });
+
+        addToTable.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        addToTable.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/addlite.png"))); // NOI18N
+        addToTable.setMaximumSize(new java.awt.Dimension(180, 47));
+        addToTable.setMinimumSize(new java.awt.Dimension(180, 47));
+        addToTable.setPreferredSize(new java.awt.Dimension(180, 47));
+        addToTable.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                addToTableActionPerformed(evt);
+            }
+        });
+
+        jButton5.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        jButton5.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/clearbtn.png"))); // NOI18N
+        jButton5.setMaximumSize(new java.awt.Dimension(180, 47));
+        jButton5.setMinimumSize(new java.awt.Dimension(180, 47));
+        jButton5.setPreferredSize(new java.awt.Dimension(180, 47));
+        jButton5.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton5ActionPerformed(evt);
+            }
+        });
+
+        customerbill.setModel(new javax.swing.table.DefaultTableModel(
+            new Object [][] {
+
+            },
+            new String [] {
+                "Item ID", "Item Name", "Brand", "Size", "Unit Price", "Quantity", "Total Price"
+            }
+        ));
+        customerbill.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                customerbillMouseClicked(evt);
+            }
+        });
+        jScrollPane2.setViewportView(customerbill);
+
+        jLabel4.setFont(new java.awt.Font("Tahoma", 1, 24)); // NOI18N
+        jLabel4.setText("Bill");
+
+        jLabel15.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        jLabel15.setText("Bill ID              :");
+
+        jLabel7.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        jLabel7.setText("Total Price      :");
+
+        jLabel8.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        jLabel8.setText("Discount        :");
+
+        txtbillid.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        txtbillid.setText("000");
+
+        txttot.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        txttot.setText("0000.00");
+
+        disfield.setCursor(new java.awt.Cursor(java.awt.Cursor.TEXT_CURSOR));
+        disfield.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                disfieldActionPerformed(evt);
+            }
+        });
+
+        txtnet.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        txtnet.setText("0000.00");
+
+        jLabel9.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        jLabel9.setText("Net Amount   :");
+
+        calcbill.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        calcbill.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/clcbill.png"))); // NOI18N
+        calcbill.setMaximumSize(new java.awt.Dimension(180, 47));
+        calcbill.setMinimumSize(new java.awt.Dimension(180, 47));
+        calcbill.setPreferredSize(new java.awt.Dimension(180, 47));
+        calcbill.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                calcbillActionPerformed(evt);
+            }
+        });
+
+        jButton6.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        jButton6.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/printbil.png"))); // NOI18N
+        jButton6.setMaximumSize(new java.awt.Dimension(180, 47));
+        jButton6.setMinimumSize(new java.awt.Dimension(180, 47));
+        jButton6.setPreferredSize(new java.awt.Dimension(180, 47));
+
+        btnUpdate.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        btnUpdate.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/Update.png"))); // NOI18N
+        btnUpdate.setMaximumSize(new java.awt.Dimension(180, 47));
+        btnUpdate.setMinimumSize(new java.awt.Dimension(180, 47));
+        btnUpdate.setPreferredSize(new java.awt.Dimension(180, 47));
+        btnUpdate.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnUpdateActionPerformed(evt);
+            }
+        });
+
+        btnDelete.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        btnDelete.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/Remove.png"))); // NOI18N
+        btnDelete.setMaximumSize(new java.awt.Dimension(180, 47));
+        btnDelete.setMinimumSize(new java.awt.Dimension(180, 47));
+        btnDelete.setPreferredSize(new java.awt.Dimension(180, 47));
+        btnDelete.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnDeleteActionPerformed(evt);
+            }
+        });
+
+        AddToBill.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
+        AddToBill.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/BarButtons/Set/addtobil.png"))); // NOI18N
+        AddToBill.setMaximumSize(new java.awt.Dimension(180, 47));
+        AddToBill.setMinimumSize(new java.awt.Dimension(180, 47));
+        AddToBill.setPreferredSize(new java.awt.Dimension(180, 47));
+        AddToBill.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                AddToBillActionPerformed(evt);
+            }
+        });
+
+        jLabel11.setFont(new java.awt.Font("Tahoma", 1, 20)); // NOI18N
+        jLabel11.setText("Add to Billing");
+
+        javax.swing.GroupLayout jPanel2Layout = new javax.swing.GroupLayout(jPanel2);
+        jPanel2.setLayout(jPanel2Layout);
+        jPanel2Layout.setHorizontalGroup(
+            jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel2Layout.createSequentialGroup()
+                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(jPanel2Layout.createSequentialGroup()
+                        .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(jButton5, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(addToTable, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(531, 531, 531))
+                    .addGroup(jPanel2Layout.createSequentialGroup()
+                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addGroup(jPanel2Layout.createSequentialGroup()
+                                .addGap(38, 38, 38)
+                                .addComponent(jLabel3, javax.swing.GroupLayout.PREFERRED_SIZE, 125, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(18, 18, 18)
+                                .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, 190, javax.swing.GroupLayout.PREFERRED_SIZE))
+                            .addGroup(jPanel2Layout.createSequentialGroup()
+                                .addGap(124, 124, 124)
+                                .addComponent(chckbtn, javax.swing.GroupLayout.PREFERRED_SIZE, 130, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)))
+                .addComponent(jLabel4, javax.swing.GroupLayout.PREFERRED_SIZE, 118, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(105, 105, 105))
+            .addGroup(jPanel2Layout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(jPanel2Layout.createSequentialGroup()
+                        .addComponent(jScrollPane2, javax.swing.GroupLayout.PREFERRED_SIZE, 585, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addGroup(jPanel2Layout.createSequentialGroup()
+                                .addGap(18, 18, 18)
+                                .addComponent(AddToBill, javax.swing.GroupLayout.PREFERRED_SIZE, 146, javax.swing.GroupLayout.PREFERRED_SIZE))
+                            .addGroup(jPanel2Layout.createSequentialGroup()
+                                .addGap(28, 28, 28)
+                                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                                    .addComponent(btnDelete, javax.swing.GroupLayout.PREFERRED_SIZE, 125, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                    .addComponent(btnUpdate, javax.swing.GroupLayout.PREFERRED_SIZE, 125, javax.swing.GroupLayout.PREFERRED_SIZE))))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                            .addGroup(jPanel2Layout.createSequentialGroup()
+                                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
+                                    .addGroup(jPanel2Layout.createSequentialGroup()
+                                        .addComponent(jLabel7, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 88, Short.MAX_VALUE)
+                                        .addComponent(txttot))
+                                    .addGroup(jPanel2Layout.createSequentialGroup()
+                                        .addComponent(jLabel15)
+                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                        .addComponent(txtbillid))
+                                    .addGroup(jPanel2Layout.createSequentialGroup()
+                                        .addComponent(jLabel9, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                        .addComponent(txtnet))
+                                    .addGroup(javax.swing.GroupLayout.Alignment.LEADING, jPanel2Layout.createSequentialGroup()
+                                        .addComponent(jLabel8, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                            .addComponent(jSeparator4)
+                                            .addComponent(disfield)
+                                            .addComponent(jSeparator6)
+                                            .addComponent(jSeparator7))))
+                                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                            .addGroup(jPanel2Layout.createSequentialGroup()
+                                .addComponent(calcbill, javax.swing.GroupLayout.PREFERRED_SIZE, 147, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                .addComponent(jButton6, javax.swing.GroupLayout.PREFERRED_SIZE, 124, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(25, 25, 25))))
+                    .addGroup(jPanel2Layout.createSequentialGroup()
+                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addGroup(jPanel2Layout.createSequentialGroup()
+                                .addGap(0, 41, Short.MAX_VALUE)
+                                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel2Layout.createSequentialGroup()
+                                        .addComponent(jLabel5, javax.swing.GroupLayout.PREFERRED_SIZE, 76, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                        .addGap(19, 19, 19)
+                                        .addComponent(iid, javax.swing.GroupLayout.PREFERRED_SIZE, 232, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel2Layout.createSequentialGroup()
+                                        .addComponent(jLabel6, javax.swing.GroupLayout.PREFERRED_SIZE, 79, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                        .addGap(18, 18, 18)
+                                        .addComponent(quanty, javax.swing.GroupLayout.PREFERRED_SIZE, 232, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                                .addGap(67, 67, 67))
+                            .addGroup(jPanel2Layout.createSequentialGroup()
+                                .addComponent(jSeparator1)
+                                .addGap(18, 18, 18))
+                            .addGroup(jPanel2Layout.createSequentialGroup()
+                                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                    .addGroup(jPanel2Layout.createSequentialGroup()
+                                        .addGap(122, 122, 122)
+                                        .addComponent(jLabel10, javax.swing.GroupLayout.PREFERRED_SIZE, 169, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                    .addGroup(jPanel2Layout.createSequentialGroup()
+                                        .addGap(116, 116, 116)
+                                        .addComponent(jLabel11, javax.swing.GroupLayout.PREFERRED_SIZE, 169, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)))
+                        .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 642, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(36, 36, 36))))
+            .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(jPanel2Layout.createSequentialGroup()
+                    .addGap(531, 531, 531)
+                    .addComponent(temp_bar_holder)
+                    .addContainerGap(594, Short.MAX_VALUE)))
+        );
+        jPanel2Layout.setVerticalGroup(
+            jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel2Layout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                    .addGroup(jPanel2Layout.createSequentialGroup()
+                        .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 277, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(18, 18, 18)
+                        .addComponent(jLabel4))
+                    .addGroup(jPanel2Layout.createSequentialGroup()
+                        .addComponent(jLabel10)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(jLabel3, javax.swing.GroupLayout.PREFERRED_SIZE, 35, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                        .addComponent(chckbtn, javax.swing.GroupLayout.PREFERRED_SIZE, 45, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jSeparator1, javax.swing.GroupLayout.PREFERRED_SIZE, 7, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(1, 1, 1)
+                        .addComponent(jLabel11)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(iid, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(jLabel5, javax.swing.GroupLayout.PREFERRED_SIZE, 35, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addGap(11, 11, 11)
+                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(quanty, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(jLabel6, javax.swing.GroupLayout.PREFERRED_SIZE, 35, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(jButton5, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(addToTable, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE))))
+                .addGap(18, 18, 18)
+                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jScrollPane2, javax.swing.GroupLayout.PREFERRED_SIZE, 219, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGroup(jPanel2Layout.createSequentialGroup()
+                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(jLabel15)
+                            .addComponent(txtbillid))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(jLabel7, javax.swing.GroupLayout.PREFERRED_SIZE, 20, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(txttot)
+                            .addComponent(AddToBill, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addGap(0, 0, 0)
+                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(disfield, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(jLabel8, javax.swing.GroupLayout.PREFERRED_SIZE, 20, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                        .addComponent(jSeparator4, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(0, 0, 0)
+                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addGroup(jPanel2Layout.createSequentialGroup()
+                                .addComponent(btnUpdate, javax.swing.GroupLayout.PREFERRED_SIZE, 35, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(18, 18, 18)
+                                .addComponent(btnDelete, javax.swing.GroupLayout.PREFERRED_SIZE, 35, javax.swing.GroupLayout.PREFERRED_SIZE))
+                            .addGroup(jPanel2Layout.createSequentialGroup()
+                                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                                    .addComponent(txtnet)
+                                    .addComponent(jLabel9, javax.swing.GroupLayout.PREFERRED_SIZE, 20, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addComponent(jSeparator6, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addComponent(jSeparator7, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(34, 34, 34)
+                                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                                    .addComponent(calcbill, javax.swing.GroupLayout.PREFERRED_SIZE, 30, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                    .addComponent(jButton6, javax.swing.GroupLayout.PREFERRED_SIZE, 30, javax.swing.GroupLayout.PREFERRED_SIZE))))))
+                .addContainerGap(18, Short.MAX_VALUE))
+            .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(jPanel2Layout.createSequentialGroup()
+                    .addGap(294, 294, 294)
+                    .addComponent(temp_bar_holder)
+                    .addContainerGap(296, Short.MAX_VALUE)))
+        );
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
+        this.setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addComponent(Left, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap())
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(Left, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGroup(layout.createSequentialGroup()
+                        .addContainerGap()
+                        .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                .addContainerGap())
+        );
+    }// </editor-fold>//GEN-END:initComponents
+
+    protected int bid = 1;
+    private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_jButton1ActionPerformed
+
+    private void disfieldActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_disfieldActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_disfieldActionPerformed
+
+    private void customerbillMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_customerbillMouseClicked
+        // TODO add your handling code here:
+        DefaultTableModel tablecus = (DefaultTableModel) customerbill.getModel();
+
+        String id = tablecus.getValueAt(customerbill.getSelectedRow(), 0).toString();
+        temp_bar_holder.setText(id);
+        System.out.println(id);
+    }//GEN-LAST:event_customerbillMouseClicked
+
+    private void jButton5ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton5ActionPerformed
+        // TODO add your handling code here:
+        quanty.setText("");
+        iid.setText("");
+    }//GEN-LAST:event_jButton5ActionPerformed
+
+    private void addToTableActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_addToTableActionPerformed
+        // TODO add your handling code here:
+        try{
+            double id = Double.parseDouble(iid.getText());
+            double qun = Double.parseDouble(quanty.getText());
+            DBHelper b = new DBHelper();
+            b.AddTocustomerbill(id,qun,bid);
+            ShowToCustomerBill(bid);
+        }catch(Exception e){
+            JOptionPane.showMessageDialog(this,"Error input","Error",JOptionPane.ERROR_MESSAGE);
+        }
+    }//GEN-LAST:event_addToTableActionPerformed
+
+    private void quantyActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_quantyActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_quantyActionPerformed
+
+    private void iidActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_iidActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_iidActionPerformed
+
+    private void chckbtnActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_chckbtnActionPerformed
+        // TODO add your handling code here:
+        String BID = jTextField1.getText();
+        DBHelper obj3 = new DBHelper();
+        boolean x = obj3.checkCustomer(BID);
+
+        if(x == true){
+            String name = obj3.getcustomerName(BID);
+            System.out.println(name);
+            JOptionPane.showMessageDialog(this,name,"Customer Found",JOptionPane.INFORMATION_MESSAGE);
+            bid++;
+        }else {
+            JOptionPane.showMessageDialog(this,"Customer Not Found","Error",JOptionPane.ERROR_MESSAGE);
+        }
+    }//GEN-LAST:event_chckbtnActionPerformed
+
+    private void jTextField1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jTextField1ActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_jTextField1ActionPerformed
+
+    private void calcbillActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_calcbillActionPerformed
+        // TODO add your handling code here:
+        Double tot = 0.00, dis =0.00;
+        try{
+            tot = Double.parseDouble(txttot.getText());
+            dis = Double.parseDouble(disfield.getText());
+        }catch(NumberFormatException e){
+            JOptionPane.showMessageDialog(this,"Input Number Value","Error",JOptionPane.ERROR_MESSAGE);
+        }
+        DBHelper obj7 = new DBHelper();
+        double neta = tot - dis;
+        txtnet.setText(String.format("%.2f", neta));
+        obj7.AddToBarBill(bid, tot, dis, neta, bid);
+
+    }//GEN-LAST:event_calcbillActionPerformed
+
+    private void btnUpdateActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnUpdateActionPerformed
+        // TODO add your handling code here:
+        if(temp_bar_holder.getText().equals("")){
+            JOptionPane.showMessageDialog(this,"Select a row","Not Found",JOptionPane.INFORMATION_MESSAGE);
+        }
+        try{
+            int ivid = Integer.parseInt(temp_bar_holder.getText());
+            DBHelper obj6 =new DBHelper();
+            //int qun = obj6.getQunFromcustomerbill(ivid);
+            int newqun = Integer.parseInt(JOptionPane.showInputDialog("Input new quntity here"));
+            obj6.updateFromcustomerbill(ivid,newqun);
+            ShowToCustomerBill(bid);
+        }catch(Exception e){
+            e.printStackTrace();
+        }
+
+    }//GEN-LAST:event_btnUpdateActionPerformed
+
+    private void btnDeleteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnDeleteActionPerformed
+        // TODO add your handling code here:
+        try{
+            int ivid = Integer.parseInt(temp_bar_holder.getText());
+            DBHelper obj5 =new DBHelper();
+            int x = obj5.deleteFromcustomerbill(ivid);
+            System.out.println(x);
+            if(x == 1){
+                JOptionPane.showMessageDialog(this,"Delete Success..!","Done",JOptionPane.INFORMATION_MESSAGE);
+            }else{
+                JOptionPane.showMessageDialog(this,"Delete Failed..!","Error",JOptionPane.ERROR_MESSAGE);
+            }
+            ShowToCustomerBill(bid);
+        }catch(Exception e){
+            JOptionPane.showMessageDialog(this,"Select a row","Not Found",JOptionPane.INFORMATION_MESSAGE);
+        }
+        int ivid = Integer.parseInt(temp_bar_holder.getText());
+        DBHelper obj5 =new DBHelper();
+        obj5.deleteFromcustomerbill(ivid);
+        ShowToCustomerBill(bid);
+    }//GEN-LAST:event_btnDeleteActionPerformed
+
+    private void AddToBillActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_AddToBillActionPerformed
+        // TODO add your handling code here:
+        DBHelper obj4 =new DBHelper();
+        double tot = obj4.CalcFromcustomerbill(bid);
+        txtbillid.setText(""+bid);
+        txttot.setText(String.format("%.2f", tot));
+    }//GEN-LAST:event_AddToBillActionPerformed
+
+
+    public void ShowInventory(){
+        DBHelper obj1 = new DBHelper();
+        ResultSet rs =  obj1.getInventory();
+        liquortable.setModel(DbUtils.resultSetToTableModel(rs));
+               
+    }
+    public void  ShowToCustomerBill(int bid){
+        DBHelper obj1 = new DBHelper();
+        ResultSet rs =  obj1.getFromcustomerbill(bid);
+        customerbill.setModel(DbUtils.resultSetToTableModel(rs));
+    }
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JButton AddToBill;
+    private javax.swing.JPanel Left;
+    private javax.swing.JButton addToTable;
+    private javax.swing.JButton btnDelete;
+    private javax.swing.JButton btnUpdate;
+    private javax.swing.JButton calcbill;
+    private javax.swing.JButton chckbtn;
+    private javax.swing.JTable customerbill;
+    private javax.swing.JTextField disfield;
+    private javax.swing.JTextField iid;
+    private javax.swing.JButton jButton1;
+    private javax.swing.JButton jButton2;
+    private javax.swing.JButton jButton3;
+    private javax.swing.JButton jButton4;
+    private javax.swing.JButton jButton5;
+    private javax.swing.JButton jButton6;
+    private javax.swing.JLabel jLabel10;
+    private javax.swing.JLabel jLabel11;
+    private javax.swing.JLabel jLabel15;
+    private javax.swing.JLabel jLabel3;
+    private javax.swing.JLabel jLabel4;
+    private javax.swing.JLabel jLabel5;
+    private javax.swing.JLabel jLabel6;
+    private javax.swing.JLabel jLabel7;
+    private javax.swing.JLabel jLabel8;
+    private javax.swing.JLabel jLabel9;
+    private javax.swing.JPanel jPanel2;
+    private javax.swing.JScrollPane jScrollPane1;
+    private javax.swing.JScrollPane jScrollPane2;
+    private javax.swing.JSeparator jSeparator1;
+    private javax.swing.JSeparator jSeparator4;
+    private javax.swing.JSeparator jSeparator6;
+    private javax.swing.JSeparator jSeparator7;
+    private javax.swing.JTextField jTextField1;
+    private javax.swing.JTable liquortable;
+    private javax.swing.JTextField quanty;
+    private javax.swing.JLabel temp_bar_holder;
+    private javax.swing.JLabel txtbillid;
+    private javax.swing.JLabel txtnet;
+    private javax.swing.JLabel txttot;
+    // End of variables declaration//GEN-END:variables
+}
