+++ src/Project/MainMenuPanel.java
@@ -33,7 +33,7 @@ public class MainMenuPanel extends javax.swing.JPanel {
         jPanel1 = new javax.swing.JPanel();
         dashboard = new javax.swing.JLabel();
         HRMannagement = new javax.swing.JLabel();
-        jLabel2 = new javax.swing.JLabel();
+        BarBilling = new javax.swing.JLabel();
         ReceptionManagement = new javax.swing.JLabel();
         KitchenManagement = new javax.swing.JLabel();
         jLabel5 = new javax.swing.JLabel();
@@ -60,13 +60,13 @@ public class MainMenuPanel extends javax.swing.JPanel {
         });
         jPanel1.add(HRMannagement, new org.netbeans.lib.awtextra.AbsoluteConstraints(460, 160, -1, 163));
 
-        jLabel2.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/Button-3.png"))); // NOI18N
-        jLabel2.addMouseListener(new java.awt.event.MouseAdapter() {
+        BarBilling.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/Button-3.png"))); // NOI18N
+        BarBilling.addMouseListener(new java.awt.event.MouseAdapter() {
             public void mouseClicked(java.awt.event.MouseEvent evt) {
-                jLabel2barMouceClicked(evt);
+                BarBillingbarMouceClicked(evt);
             }
         });
-        jPanel1.add(jLabel2, new org.netbeans.lib.awtextra.AbsoluteConstraints(637, 369, -1, -1));
+        jPanel1.add(BarBilling, new org.netbeans.lib.awtextra.AbsoluteConstraints(637, 369, -1, -1));
 
         ReceptionManagement.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/Button-7.png"))); // NOI18N
         ReceptionManagement.setPreferredSize(new java.awt.Dimension(100, 100));
@@ -139,13 +139,11 @@ public class MainMenuPanel extends javax.swing.JPanel {
         h1.setVisible(true);
     }//GEN-LAST:event_HRMannagementMouseClicked
 
-    private void jLabel2barMouceClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jLabel2barMouceClicked
+    private void BarBillingbarMouceClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_BarBillingbarMouceClicked
         // TODO add your handling code here:
-        Billing b1 = new Billing();
-        this.setVisible(false);
-        b1.setVisible(true);
+        changeFunction("Bar");
 
-    }//GEN-LAST:event_jLabel2barMouceClicked
+    }//GEN-LAST:event_BarBillingbarMouceClicked
 
     private void ReceptionManagementMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_ReceptionManagementMouseClicked
         //Navigate to the Reception Management.
@@ -201,6 +199,7 @@ public class MainMenuPanel extends javax.swing.JPanel {
 
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JLabel BarBilling;
     private javax.swing.JLabel HRMannagement;
     private javax.swing.JLabel InventoryBtn;
     private javax.swing.JLabel KitchenManagement;
@@ -208,7 +207,6 @@ public class MainMenuPanel extends javax.swing.JPanel {
     private javax.swing.JLabel TransactionManagement;
     private javax.swing.JLabel borderline;
     private javax.swing.JLabel dashboard;
-    private javax.swing.JLabel jLabel2;
     private javax.swing.JLabel jLabel5;
     private javax.swing.JLabel jLabel7;
     private javax.swing.JPanel jPanel1;
