+++ src/ProjectDB/DbClass.java
@@ -26,7 +26,7 @@ public class DbClass {
         try {
             Class.forName("com.mysql.cj.jdbc.Driver");
           
-            Connection con = DriverManager.getConnection("jdbc:mysql://localhost:3306/itpfinaldb","root","password");
+            Connection con = DriverManager.getConnection("jdbc:mysql://localhost:3306/itpfinaldb","root","nc@12345");
             //When clone the project, Make sure to change the username and password accroding to your server.
             //This is for testing purposes.
             //Comment
